+++ src/core/task-manager.js
@@ -22,7 +22,7 @@ class TaskManager {
    * @private
    */
   _generateTaskId() {
-    return `task_${Date.now()}_${Math.random().toString(36).substr(2, 9)}`;
+    return `task_${Date.now()}_${Math.random().toString(36).substring(2, 11)}`;
   }
 
   /**